requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
numpy<2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
requests>=2.31.0
tqdm>=4.66.0
gputil>=1.4.0
GitPython>=3.1.30